Extract repo-icon detection into dedicated module, fix base dir resolution
Move icon scanning logic out of git-context.ts into shared/repo-icon.ts
to keep git probing focused on git metadata. Fix the repo-icon endpoint
to resolve icon paths relative to gitContext.worktreePath (the checkout
root where the scan ran) instead of agent.cwd, which may be a nested
subdirectory.
